+++ src/Concerns/ManagesStaleReads.php
@@ -32,9 +32,7 @@ trait ManagesStaleReads
     {
         return $this->run($query, $bindings, function ($query, $bindings) use ($timestampBound) {
             if ($this->pretending()) {
-                return call_user_func(static function() {
-                    yield from [];
-                });
+                return (static fn() => yield from [])();
             }
 
             $options = ['parameters' => $this->prepareBindings($bindings)];