+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # v6.0.0 [Not released Yet]
 
 Changed
-- [Breaking] Match `Quer\Builder::forceIndex()` behavior with laravel's (`forceIndex` property no longer exists). (#114)
+- [Breaking] Match `Query\Builder::forceIndex()` behavior with laravel's (`forceIndex` property no longer exists). (#114)
 
 # v5.2.0 [Not released Yet]