+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Added
 Changed
 - `Connection::waitForOperation` and `Connection::isDoneOperation` has been removed. (#99)
 - Update `export-ignore` entries in `.gitattributes` (#104)
+- Stop using `call_user_func` (#121)
 
 Fixed
 - Transaction state was not being cleared if rolled back failed. (#107)